+++ src/config/command.rs
@@ -72,6 +72,7 @@ pub struct SuperviseInfo {
 
     // Supervise
     pub mode: SuperviseMode,
+    pub kill_unresponsive_after: Option<u32>,
     pub children: BTreeMap<String, ChildCommand>,
 }
 
@@ -234,6 +235,10 @@ pub fn command_validator<'a>() -> Box<V::Validator + 'a> {
             key_element: Box::new(V::Scalar {
                 ..Default::default()}) as Box<V::Validator>,
             value_element: subcommand_validator(),
+            .. Default::default()}) as Box<V::Validator>),
+        ("kill_unresponsive_after".to_string(), Box::new(V::Scalar {
+            default: None,
+            optional: true,
             .. Default::default()}) as Box<V::Validator>),
         );
     supervise_members.extend(command_fields().into_iter());